requirements.txt
requests==2.24.0
torch==1.4.0
numpy==1.20.1
Flask==1.1.2
boto3==1.16.9
botocore==1.19.9
websocket_client==0.57.0
aws-cdk.core==1.85.0
aws-cdk.aws-ec2==1.85.0
aws-cdk.aws-rds==1.85.0
aws-cdk.aws-ecs==1.85.0
aws-cdk.aws-logs==1.85.0
aws-cdk.aws-ecs-patterns==1.85.0
aws-cdk.aws-elasticloadbalancingv2==1.85.0